Move OctoprintConfigEntry and OctoprintData to coordinator module
